(#281) Prevents Creation Of Unrequired Self-Signed Certificate
The Chocolatey-Management-Service package generates a self-signed certificate to use if no thumbprint is specified as a parameter.

Regardless of if a certificate was specified, we were installing it without a thumbprint and then setting the certificate later.

This resulted in an unused self-signed certificate being generated and stored.

This change ensures the package gets the appropriate parameter in order to not generate unneeded certificate(s).